+++ src/nodes/math/Hash.js
